Moved PdfHandling Classes into an own Project
