+++ PdfHandling/PreviewablePdfDocument.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using System.Windows.Media;
 using PdfiumLight;
 
-namespace ZebraDesktop
+namespace Zebra.PdfHandling
 {
     public class PreviewablePdfDocument : PdfDocument
     {
