--- ZebraDesktop/PdfHandling/Converter.cs
@@ -1,27 +1,0 @@
-﻿using System.Drawing.Imaging;
-using System.IO;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-
-namespace ZebraDesktop
-{
-    public static class Converter
-    {
-        public static ImageSource ConvertToImageSource(this System.Drawing.Image value)
-        {
-            using (var ms = new MemoryStream())
-            {
-                value.Save(ms, ImageFormat.Bmp);
-                ms.Seek(0, SeekOrigin.Begin);
-
-                var bitmapImage = new BitmapImage();
-                bitmapImage.BeginInit();
-                bitmapImage.CacheOption = BitmapCacheOption.OnLoad;
-                bitmapImage.StreamSource = ms;
-                bitmapImage.EndInit();
-
-                return bitmapImage;
-            }
-        }
-    }
-}
